requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
langchain
python-dotenv
pydantic
faiss-cpu